Add SwiftNumber in Read Dto and update DB Script

+++ EDocument_Data/DTOs/Requests/FuelOilInvoiceRequest/ApproveFuelOilInvoiceRequestDto.cs
@@ -9,7 +9,7 @@ namespace EDocument_Data.DTOs.Requests.FuelOilInvoiceRequest
         [Required]
         public long RequestId { get; set; }
         [Required]
-        public long SwiftNumber { get; set; }
+        public int SwiftNumber { get; set; }
         public string? ReviewedNotes { get; set; }
     }
 }
